+++ server/index.js
@@ -1,8 +1,8 @@
 const express = require("express");
 var bodyParser = require('body-parser');
-const authRouter = require("./auth/router");
-const userRouter = require("./user/router");
-const clubRouter = require("./club/router");
+const authRouter = require("./auth/AuthRouter");
+const userRouter = require("./user/UserRouter");
+const clubRouter = require("./club/ClubRouter");
 
 const app = express();
 app.use(bodyParser.json())
